+++ src/FE/Driver.h
@@ -47,9 +47,9 @@ class Driver :
 	public AST::Build::Builder
 {
 public:
-	static AST::Ptr <const AST::Root> parse (const std::string& file, std::istream& yyin, bool anonymous_deprecated)
+	static AST::Ptr <const AST::Root> parse (const std::string& file, std::istream& yyin, bool anonymous_deprecated, std::ostream& err_out)
 	{
-		Driver drv (file, yyin, anonymous_deprecated);
+		Driver drv (file, yyin, anonymous_deprecated, err_out);
 		return drv.parse ();
 	}
 
@@ -69,7 +69,7 @@ private:
 		return *this;
 	}
 
-	Driver (const std::string& file, std::istream& yyin, bool anonymous_deprecated);
+	Driver (const std::string& file, std::istream& yyin, bool anonymous_deprecated, std::ostream& err_out);
 
 	AST::Ptr <const AST::Root> parse ()
 	{