+++ src/include/IDL_FrontEnd.h
@@ -29,6 +29,7 @@
 #pragma once
 
 #include <string>
+#include <iostream>
 #include <assert.h>
 
 namespace AST {
@@ -120,14 +121,15 @@ protected:
 	static const unsigned FLAG_DEPRECATE_ANONYMOUS_TYPES = 1;
 
 protected:
-
 	/// Constructor
 	/// 
 	/// \param flags The flags.
 	///              Currently only IDL_FrontEnd::FLAG_DEPRECATE_ANONYMOUS_TYPES is supported.
-	IDL_FrontEnd (unsigned flags = 0) :
+	/// \param err_out Compiler messages output stream. Default is std::cerr.
+	IDL_FrontEnd (unsigned flags = 0, std::ostream& err_out = std::cerr) :
 		flags_ (flags),
-		arguments_ (nullptr)
+		arguments_ (nullptr),
+		err_out_ (err_out)
 	{}
 
 	/// Parse command line parameter.
@@ -164,6 +166,12 @@ protected:
 	///                           Other exceptions will cause the compilation interruption.
 	virtual void generate_code (const ::AST::Root& tree) = 0;
 
+	/// \returns Compiler messages output stream.
+	std::ostream& err_out () const noexcept
+	{
+		return err_out_;
+	}
+
 private:
 	bool compile (const std::string& file);
 
@@ -171,6 +179,7 @@ private:
 	unsigned flags_;
 	struct Arguments;
 	Arguments* arguments_;
+	std::ostream& err_out_;
 };
 
 /// \example IDL_Print.cpp