+++ changelog.md
@@ -4,6 +4,8 @@
 ## Changed
  - Allow native type in the exception list, as it required by the standard.
  - Added class BE::MessageOut for the convenient compiler messages output.
+ - Messages are redirected to `std::cerr` by default, instead of `std::cout`.
+ - Added parameter `std::ostream& err_out` to `IDL_FrontEnd` constructor.
  
 ## 2.1.8 - 2022-08-13
 ### Fixed